requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==25.1.0
bleach==6.2.0
bokeh==3.6.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
frozenlist==1.5.0
idna==3.10
Jinja2==3.1.5
linkify-it-py==2.0.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
multidict==6.1.0
numpy==2.2.2
packaging==24.2
pandas==2.2.3
panel==1.6.0
param==2.2.0
pillow==11.1.0
prompt_toolkit==3.0.50
propcache==0.2.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pyviz_comms==3.0.4
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
setuptools==75.8.0
six==1.17.0
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uc-micro-py==1.0.3
urllib3==2.3.0
wcwidth==0.2.13
webencodings==0.5.1

xyzservices==2025.1.0
yarl==1.18.3
